requirements.txt
fastapi==0.136.1
uvicorn==0.46.0
chromadb==1.5.8
openai==2.28.0
python-dotenv==1.2.2
PyPDF2==3.0.1
python-docx==0.8.11
pydantic==2.12.5
python-multipart==0.0.26
